added more

+++ src/resolvers/Mutation.js
@@ -59,9 +59,12 @@ const Mutation = {
             }, 'izwebishot!')
         };
     },
-    async deleteUser(parent, args, { prisma }, info) {
-       const userExists = await prisma.exists.User({
-           id: args.id
+    async deleteUser(parent, args, { prisma, request }, info) {
+       
+        const userId = getUserId(request);
+       
+        const userExists = await prisma.exists.User({
+           id: userId
        });
 
        if(!userExists) {
@@ -70,14 +73,17 @@ const Mutation = {
 
        return prisma.mutation.deleteUser({
            where: {
-               id: args.id
+               id: userId
            }
        });
     },
-    async updateUser(parent, args, { prisma }, info) {
+    async updateUser(parent, args, { prisma, request }, info) {
+        
+        const userId = getUserId(request);
+        
         return prisma.mutation.updateUser({
             where: {
-                id: args.id
+                id: userId
             },
             data: args.data
         }, info);
@@ -99,14 +105,43 @@ const Mutation = {
             }
          }, info);
     },
-    async deletePost(parent, args, { prisma }, info) {
+    async deletePost(parent, args, { prisma, request }, info) {
+        
+        const userId = getUserId(request);
+
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        });
+
+        if(!postExists) {
+            throw new Error('Unable to delete post');
+        }
+        
         return prisma.mutation.deletePost({
             where: {
                 id: args.id
             }
         });
     },
-    async updatePost(parent, args, { prisma }, info) {
+    async updatePost(parent, args, { prisma, request }, info) {
+        
+        const userId = getUserId(request);
+        const postExists = await prisma.exists.Post({
+           
+                id: args.id,
+                author: {
+                    id: userId
+                }
+            
+        });
+
+        if(!postExists){
+            throw new Error('Unable to update post');
+        }
+        
         return prisma.mutation.updatePost({
             where: {
                 id: args.id
